make Task an abstract class
